refactor: rename image preprocessing file
