+++ backend/src/gee/image_preprocessing.py
