+++ backend/src/service.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone, timedelta
 import pandas as pd
 
 from src.constants import AggregationMethod, LocationPolygon, TemporalResolution
-from src.gee.index import get_preprocessed_imagery
+from src.gee.image_preprocessing import get_preprocessed_imagery
 from src.gee.ndvi import get_ndvi_info
 from src.gee.ndvi_cache import ndvi_daily_cache
 from typing import List, Dict, Union
